requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
passlib[bcrypt]
PyJWT
scikit-learn
joblib
numpy
python-multipart
